[REEF-423]: Move Streaming Codec classes from Network to Wake layer
This addressed the issue by
  * actually moving the StreamingCodec to Wake layer

JIRA: [REEF-423](https://issues.apache.org/jira/browse/REEF-423)
