+++ lang/cs/Org.Apache.REEF.Network/Group/Pipelining/StreamingPipelineMessageCodec.cs
@@ -21,7 +21,7 @@ using Org.Apache.REEF.Tang.Annotations;
 using Org.Apache.REEF.Wake.Remote;
 using System.Threading;
 using System.Threading.Tasks;
-using Org.Apache.REEF.Network.StreamingCodec;
+using Org.Apache.REEF.Wake.StreamingCodec;
 
 namespace Org.Apache.REEF.Network.Group.Pipelining
 {
